+++ pages/c.tsx
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export default () => <div>c</div>
